MEN-2232: Write update module for doing container setup
Simple update module to setup Docker containers to run on target.

It uses the meta-data field from Mender Artifact to send the image
hashes desired to be running in the target. Supoprts rollback.

Added also a script to generate the artifacts with the meta-data format
that the update module requires.

Changelog: Title

Signed-off-by: Lluis Campos <lluis.campos@northern.tech>