chore: Decouple `UpdateModule` from artifact parser.
This is important because after a reboot, we don't have an artifact
anymore, but we still need to construct an instance of the object to
carry out the remaining Update Module states.

The artifact references have been moved to the function call arguments
of the functions that need them instead.

Signed-off-by: Kristian Amlie <kristian.amlie@northern.tech>