chore: Implement AsyncReader interface for consuming HTTP body.
This is necessary for more advanced consumption of the body, like
reading an artifact. One of the key motivators is that you can report
errors through an AsyncReader, which you cannot do with a Writer
interface (the writer can report errors back to you, but you cannot
report errors to it).

The old writer interface is still kept, and both can be used if
necessary (though this is probably rare).

Some key things to point out in the change:

* If an AsyncReader has been created, then the progress of the
  download is driven by calling AsyncRead, instead of automatically.

* With calling of many handlers, it started to become tricky to track
  what happened inside a handler. Any of them can cancel a request and
  start a new one. Therefore, create the `cancelled_` bool member,
  which we replace with a new bool every time we cancel something. If
  `cancelled_` is true after the calling of any handler, we just
  return straight away.

* The Boost Beast `is_done()` call doesn't seem to work correctly in
  all cases. It seems it can return true before we have read the whole
  body, if it has internally parsed it and it's sitting in its
  buffer. This makes it impossible to tell when the end has been
  reached, so we start tracking the number of bytes read instead, and
  compare it to the expected content-length.

Signed-off-by: Kristian Amlie <kristian.amlie@northern.tech>