refac: Fix confusing parsing interface and associated code.
This was my own mistake, but it's an easy one to
make. `CmdlineOptionsIterator` takes two arguments, options with
values, and options without values. However the two functions
`CommandOptsSetWithValue` and `CommandOptsSetWithoutValue` will filter
and construct the correct lists from one single list of options, but
it's easy to conclude that the options need to be kept in separate
lists. This was made even more confusing by the call sites not passing
in this one list consistently, but instead passing it in depending on
whether that particular command has arguments or not. This is not
necessary, since the two aforementioned functions would automagically
construct the correct (empty) lists based on filtering.

Simplify this, and give the constructor one argument instead, the full
list of options, and then it can itself do the filtering, so that it
doesn't look like we are using two different lists from call sites.

This removes the need for the `options_w_values` list.

Signed-off-by: Kristian Amlie <kristian.amlie@northern.tech>